refactor(query): extract post-sampling hooks into runner/hooks.rs
